requirements.txt
asyncpg
discord.py>=1.0.0a1461
uvloop
pillow
PyYAML
jishaku
yarl<1.2
